requirements.txt
neo4j>=5.14.0
fastmcp>=2.0,<3
python-dotenv>=1.0.0
